# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Innoscripta API",
        "version": "0.1"
    },
    "paths": {
        "/api/article/{id}": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Get a one article",
                "description": "Retrieve one article",
                "operationId": "eeb1c815391a2e3111d59f6dcf868b37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parameter for get article id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "data": {
                                            "id": 3,
                                            "title": "Dr.",
                                            "source": "David",
                                            "url": "http://www.corwin.com/dolore-velit-blanditiis-necessitatibus-atque-sed-cum-voluptatem.html",
                                            "date": "2000-06-27 11:07:51",
                                            "category": "development",
                                            "created_at": "2024-12-16T20:11:09.000000Z",
                                            "updated_at": "2024-12-16T20:11:09.000000Z"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/article": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Get a list of articles",
                "description": "Retrieve of list of articles",
                "operationId": "07db2b26f66d03844900e453afc8623b",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Search for category",
                        "required": false
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Search for source",
                        "required": false
                    },
                    {
                        "name": "created_at",
                        "in": "query",
                        "description": "Search for created_at",
                        "required": false
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search for keyword",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "current_page": 1,
                                        "data": [
                                            {
                                                "id": 3,
                                                "title": "Dr.",
                                                "source": "David",
                                                "url": "http://www.corwin.com/dolore-velit-blanditiis-necessitatibus-atque-sed-cum-voluptatem.html",
                                                "date": "2000-06-27 11:07:51",
                                                "category": "development",
                                                "created_at": "2024-12-16T20:11:09.000000Z",
                                                "updated_at": "2024-12-16T20:11:09.000000Z"
                                            }
                                        ],
                                        "first_page_url": "http://127.0.0.1:8000/api/article?page=1",
                                        "from": 1,
                                        "last_page": 1,
                                        "last_page_url": "http://127.0.0.1:8000/api/article?page=1",
                                        "links": [
                                            {
                                                "url": null,
                                                "label": "&laquo; Previous",
                                                "active": false
                                            },
                                            {
                                                "url": "http://127.0.0.1:8000/api/article?page=1",
                                                "label": "1",
                                                "active": true
                                            },
                                            {
                                                "url": null,
                                                "label": "Next &raquo;",
                                                "active": false
                                            }
                                        ],
                                        "next_page_url": null,
                                        "path": "http://127.0.0.1:8000/api/article",
                                        "per_page": 15,
                                        "prev_page_url": null,
                                        "to": 1,
                                        "total": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}